requirements.txt
Django~=3.2.3
django-easy-maps
django-crispy-forms
